finalProject/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
cffi==1.14.5
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.1
django-bootstrap-form==3.4
django-filter==2.4.0
django-import-export==2.5.0
django-right-to-left==0.1.1
django-widget-tweaks==1.4.8
dnspython==2.1.0
email-validator==1.1.2
et-xmlfile==1.1.0
idna==3.1
lml==0.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pycodestyle==2.7.0
pycparser==2.20
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
PyMySQL==1.0.2
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
validate-email==1.3
xlrd==1.2.0
xlwt==1.3.0
